Move CreateRental to new file addrental.
